feature(vscode) Bridging files over to VSCode from editor.
