+++ editor/src/core/vscode/common
@@ -1,0 +1,1 @@
+../../../../utopia-vscode-common
